backend/models/requirements.txt
# Model dependencies for Invoice Extractor
# Install with: pip install -r models/requirements.txt

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Computer Vision models
opencv-python>=4.7.0
Pillow>=9.0.0
albumentations>=1.3.0

# YOLO
ultralytics>=8.0.0

# Transformers (BERT, LayoutLM)
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0

# DocTR (Document Text Recognition)
python-doctr[torch]>=0.6.0

# Table detection
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Configuration and utilities
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.65.0

# Model downloading
gdown>=4.7.0  # For Google Drive
boto3>=1.26.0  # For AWS S3 (optional)

# Optimization
onnx>=1.14.0  # For ONNX export (optional)
onnxruntime>=1.15.0  # For ONNX inference (optional)

# Additional OCR engines (optional)
pytesseract>=0.3.10
easyocr>=1.7.0

# GPU acceleration (optional, install separately based on hardware)
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# For CPU only: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu